Route all print requests through one function
This commit does not change existing behavior, the refactor enables the
implementation of a custom print handler in a later commit.

The viewer code should get the opportunity to handle the print request
before the mozPrintCallback polyfill intercepts Ctrl+P, so the
useCapture flag in mozPrintCallback has been set to false.

I verified that changing to useCapture=false does not change the
behavior of browsers with respect to the ability to prevent the print
dialog from being opened. In all cases, with and without the useCapture
flag, the following happens when the keydown event listener uses
event.preventDefault() upon Ctrl+P:

- Chrome 31, 43: The print dialog is suppressed.
- Firefox 3.6, 40: The print dialog is suppressed.
- Opera 31: The print dialog is suppressed.
- Safari 5.1, 8: The print dialog always appears.
- IE 9, 10, 11, Edge 0.11: The print dialog always appears.